Programacion de la seccion de sugerencias

+++ application/views/monedas/list.php
@@ -149,24 +149,28 @@ if ( $gobernante=='') {
                                                 </a>
                                             </td>
 
-                                            <td>
-                                            	<!-- <button type="button" class="btn btn-info btn-view-usuario" data-toggle="modal" data-target="#modal-default" class="btn btn-info btn-view" onclick="datosusuariomonedas(<?php echo $usuario->id_catalogo_moneda;?>)" value="<?php echo $usuario->id_catalogo_moneda;?>">
-                                                        <span class="fa fa-search"></span>
-                                                </button>-->
+                                            <td>                                            
 
                                               <?php if($this->session->userdata("tipo_usuario")==1 ):?>
                                                  <a title="Editar Moneda" href="<?php echo base_url();?>monedas/edit/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-success btn-check"><span class="fa fa-pencil"></span></a> 
+                                              <?php endif; ?> 
+
+
 
                                                 <?php if ($usuario->estado=='1'): ?>
                                                     <a title="Deshabilitar Moneda" href="<?php echo base_url();?>monedas/deshabilitar/0/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-warning"><span class="fa fa-remove"></span></a> 
                                               <?php else: ?>
                                                   <a title="Habilitar Moneda" href="<?php echo base_url();?>monedas/deshabilitar/1/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-success"><span class="fa fa-check"></span></a>                                                   
-                                              <?php endif ?>                                               
+                                              <?php endif ?>  
 
-                                                <a title="Eliminar Moneda" onclick="return confirm('Estas Seguro?')" href="<?php echo base_url();?>monedas/delete/2/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-danger"><span class="fa fa-trash-o"></span></a> 
+
+                                              <?php if($this->session->userdata("tipo_usuario")==1 ):?>
+                                                  <a title="Eliminar Moneda" onclick="return confirm('Estas Seguro?')" href="<?php echo base_url();?>monedas/delete/2/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-danger"><span class="fa fa-trash-o"></span></a> 
                                                <?php endif; ?> 
 
-                                               <a title="Agregar a Colección" href="<?php echo base_url();?>collectionm/add_collection/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-success btn-check"><span class="fa fa-plus"></span></a>         
+                                               <a title="Agregar a Colección" href="<?php echo base_url();?>collectionm/add_collection/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-success btn-check"><span class="fa fa-plus"></span></a> 
+
+                                               <a title="Sugerir Cambios" href="<?php echo base_url();?>monedas/sugerir/<?php echo $usuario->id_catalogo_moneda;?>" class="btn btn-primary btn-check"><span class="fa fa-plus-circle"></span></a>              
 
                                             </td>
                                         </tr>
